+++ data/scripts/galaxy/init.lua
@@ -8,5 +8,6 @@ if onServer() then
         galaxy:addScriptOnce("data/scripts/server/background/cosmicwarbounties.lua")
         galaxy:addScriptOnce("data/scripts/server/background/cosmicwarbridgeupdate.lua")
         galaxy:addScriptOnce("data/scripts/server/background/cosmicwardiplomacy.lua")
+        galaxy:addScriptOnce("data/scripts/server/background/cosmicwarsiege_server.lua")
     end
 end